refactor: folder structure, chore: x-axis tick values for larger data
