+++ apps/app/components/analytics/table.tsx
