+++ apps/app/components/core/index.ts
@@ -1,4 +1,3 @@
-export * from "./analytics";
 export * from "./board-view";
 export * from "./list-view";
 export * from "./sidebar";
